backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.25
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.3
pydantic-settings==2.3.0
boto3==1.34.34
httpx==0.27.0
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6
jinja2==3.1.3
pandas==2.2.0
openpyxl==3.1.2
reportlab==4.0.9
python-dateutil==2.8.2
structlog==24.1.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
twilio==9.0.1
icalendar==5.0.12
